% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_stanfit.R
\name{extract_np}
\alias{extract_np}
\title{data frame containing the required diagnostic information for the
bayesplot package}
\usage{
extract_np(mfit_diag, n_iter = 1000)
}
\description{
data frame containing the required diagnostic information for the
bayesplot package
}
